refactor: REST Core-API: IR code format for toggle & sequence codes
- The | separator is used for two PRONTO toggle codes
- The + separator is used for IR sequences of 2 or more codes

This change is required for properly handle RC-5, RC-6, RC-MM (Nokia32)
toggle IR protocols with PRONTO codes.

**Breaking change**
Manually added IR code sequences must be changed to use the + separator.
Since this was a hidden feature only mentioned in the API, not many
users should be affected.

The upside is that manufacturer codesets are not affected and already
created remote entities must not be changed. Codesets with toggle
protocols will now work correctly.

Relates to unfoldedcircle/feature-and-bug-tracker#126